Update 9-8-19
updates over the weekend
