+++ Assets/Asset Landing Zone/Armory/AccessoryRack.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d4cbac75f5b4dbe4591a25bf4ef00d96
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
