+++ Assets/Asset Landing Zone/Armory/Guns/Gun Attachments/Accessories 2_0/Accessories.blend
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:aba8bdaa60b4798f9bae75650f7a185933664d3e7ce9c4e49fc096e6493c54d3
-size 8864700
+oid sha256:06ed9f65d6e5251ba81bafc2334c781276d6207a6c7b2f9c851ca0468cb5a170
+size 12270036
