+++ Assets/Asset Landing Zone/Armory/Guns/Gun Attachments/Accessories 2_0/Accessories_Barrel.fbx
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:d7b165af2f6116f8d53119d75c5ba48aa3cc5bd89bff49c65e38d2faa247acef
+size 615484
