--- Assets/Asset Landing Zone/Armory/Guns/Gun Attachments/Accessories 2_0/Accessories_Pistol.fbx
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:ebba2056bd94c1aa943ece553777864b5397f028d672607e75c43f2e9ef6a0e1
-size 151244
