+++ Assets/Asset Landing Zone/Armory/Guns/Gun Attachments/Accessories 2_0/Accessories_SIghts.fbx
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:2b42ddec324ab70759c3be93b1c55a78a1df5bd89aaf316986ca790aa8d7fed7
+oid sha256:e4d0097057f0db5ddbbe59911faff3c2b627ab8bcdf7af3fd357e3e238554101
 size 755804
