--- Assets/Asset Landing Zone/Armory/Guns/Gun Attachments/Accessories 2_0/Accessories_Side.fbx
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:983a276efef8dfc8261bf27dfab21952ee20fac556c9d99fea8ebfa5eaf61020
-size 242188
