--- Assets/Asset Landing Zone/Armory/Guns/Gun Attachments/Accessories 2_0/Accessories_Suppressor.fbx
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:d7e62b4064815d41babc962543f1f7910afa814f7100fde138b3c8d452157d15
-size 83228
