+++ Assets/Asset Landing Zone/Armory/Guns/Gun Attachments/Accessories 2_0/Accessories_Under.fbx
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:4f3272b23b48bdb6bcd7bfc73347965946275d0f31e3e384353910f6dde71724
+oid sha256:a3ea0d4a7a76bda39af1f5ccfda26ff0b3745de8b9bc2a183e25c7283e1fe7b9
 size 1081164
