+++ Assets/Asset Landing Zone/Armory/Guns/Gun Attachments/Accessories 2_0/Accessories_Utility.fbx
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:e0a0bd328ce7ac84ffbc621f91442c6e25021eef4098dd20541236fff91f37ea
+size 377788
