+++ Assets/Asset Landing Zone/Armory/Guns/Gun Attachments/Magazines/Materials/BioBattery_Base_Color.png
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:465343f35140ca4316439c0cf59122e9c9d647503903ee9159859e0f58eb287d
-size 489900
+oid sha256:288cb249254637b76faa886b9132e344b8b5b115860eaffacdf4a4d145301cc3
+size 520154
