+++ Assets/Asset Landing Zone/Armory/Guns/Gun Attachments/Magazines/Materials/BioBattery_Normal_OpenGL.png
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:f6f2386d9597114f8e160d67bf31974dfeb0fe185906b3e1c1be5c936928418e
-size 616859
+oid sha256:a27fca36a63b8182051121f7c9de278965891748dc303d399ccdc4f1e5d75be0
+size 616857
