+++ Assets/Asset Landing Zone/Armory/Guns/Gun Attachments/Magazines/Materials/BioBattery_Roughness.png
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:6315cf0a9cb263d4a1149a72537623828463e434c47493b51dc30aa97cd0129b
-size 352796
+oid sha256:81918c5a88db5d8209dbcc65037ec046983c7281eaf9e70384e7413e991eec2c
+size 380078
