+++ Assets/Asset Landing Zone/Armory/Guns/MasterBlaster/ChemBaton.fbx
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:97cf73a5875d64383146c5e50f65952db5b1848515152553d36a14d16d722fc0
-size 862300
+oid sha256:385debd7b49fa00a185c5cde67c650272606976156f903a081e30c07dbaba600
+size 834380
