+++ Assets/Asset Landing Zone/Armory/Guns/MasterBlaster/Greasepup.fbx
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:8aa0217ae8011c761b160f0b1b5088cbe47898ba84a37417e05e99e454e4f842
-size 706652
+oid sha256:d55f022640fd4d27068dbe3e4c2d478298fe57e0dd5de6e1d4bc47c299fb726b
+size 678732
