+++ Assets/Asset Landing Zone/Armory/Guns/MasterBlaster/MasterBlaster.fbx
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:318c918d72e330ab7f3d3a8cd0904994b54886378e225148dcc3b14a0fe8093d
-size 1414796
+oid sha256:4e3d7d7609608a3e90336382b4622d1103da4670a0869fb7476c13e03691ddd2
+size 1389724
