+++ Assets/Asset Landing Zone/Armory/Guns/MasterBlaster/MasterBlasterWorkInProgress5.blend
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:d76d30ab0eb5e881762fd00d5e6f4613c2f77f245060295847ccfd0b4fc624dd
-size 17481448
+oid sha256:e44c9f371e860c973132bbb6f18f80e66485d5216ea46c058582fea510f9d6a0
+size 17330652
