+++ Assets/Asset Landing Zone/Armory/Guns/MasterBlaster/StabZapper.fbx
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:cd7df302a388e38af96b58cf5596e3145492f9d8f36423273bb84382cb6cad62
-size 1541020
+oid sha256:5fc98f8f7925983eab51ebf9f9fe5ebc0f9c3558e35454b2ddb22e65a6f703ef
+size 1515948
