+++ Assets/Asset Landing Zone/Armory/Guns/Multishot/MultishotWorkInProgress.fbx
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:53faf8ab42b9f45c003c7d1ea4e28c4ee1bd04baa42307fd13f71eeee52fad5f
-size 13861596
+oid sha256:82f7ffb0eb1fd5756413776657e228d1e9a38b97806085e4e4c0f8156b81ef81
+size 13751404
