+++ Assets/Asset Landing Zone/Armory/Guns/Multishot/MultishotWorkInProgress6.blend
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:f083bad1a62a057d833abae6a622080f7c353499decdfa2293ece6093e832f0e
-size 50995884
+oid sha256:ce72c0bfc4da1b95018a43d4366e40278b193580dc817469c60b71378cf9f0da
+size 50636660
