+++ Assets/Asset Landing Zone/Armory/Guns/Muzzleloader/SplattercasterWorkInProgress.fbx
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:01bb24ca926eb6046f9dfb6f230a70c61315e286a5dde07645dc7e236d6e9922
-size 3717516
+oid sha256:e6d8633465de4090e522c2fbb18efb4358244aee72b5e1210a450dd7eda7b255
+size 3645036
