+++ Assets/Asset Landing Zone/Armory/Guns/Muzzleloader/SplattercasterWorkInProgress2.blend
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:2a43a2b587d364b6b2deafade8e7111385cd9620af3952e5203d954daeea3d2f
-size 14238516
+oid sha256:40e092212dc0d60423f29cace7928526c106891f0797efca359fb987e921f00d
+size 13939456
