+++ Assets/Asset Landing Zone/Armory/WeaponCustomizerConfigurator.cs
@@ -19,55 +19,6 @@ public class WeaponCustomizerConfigurator : MonoBehaviour
 		MeleeOversized,
 	}
 	
-	public enum AccessoryType
-	{
-		Sight,
-		Barrel,
-		Utility,
-		Under,
-	}
-	
-	public enum SightType
-	{
-		Standard,
-		Pistol_Dot,
-		Red_Dot,
-		ACOG,
-		Scope16x,
-		ScopeEnhanced,
-	}
-	
-	public enum UtilityType
-	{
-		Standard,
-		Light,
-		Pistol_Light,
-		Laser,
-		Pistol_Laser,
-		Light_Laser
-	}
-	
-	public enum BarrelType
-	{
-		Standard,
-		Suppressor,
-		Suppressor_Pistol,
-		Bayonet_Modern,
-		Bayonet_Classic,
-		Bayonet_Wedge,
-		Bayonet_Spike,
-	}
-	
-	public enum UnderType
-	{
-		Standard,
-		Foregrip,
-		Grenade,
-		Energy_Biochem,
-		Energy_Fission,
-		Chemical,
-	}
-	
 	[System.Serializable]
 	public class Part
 	{
@@ -171,17 +122,6 @@ public class WeaponCustomizerConfigurator : MonoBehaviour
 		[Tooltip("Specifies how the clip should be reloaded.")]
 		[SerializeField] public Opsive.UltimateCharacterController.Items.Actions.ShootableWeapon.ReloadClipType m_ReloadType = 
 			Opsive.UltimateCharacterController.Items.Actions.ShootableWeapon.ReloadClipType.Full;
-	}
-	
-	[System.Serializable]
-	public class AccessoryRack
-	{
-		[SerializeField]public SightType m_ActiveSight = SightType.Standard;
-		[SerializeField]public BarrelType m_ActiveBarrel = BarrelType.Standard;
-		[SerializeField]public UtilityType m_ActiveSideL = UtilityType.Standard;
-		[SerializeField]public UtilityType m_ActiveSideR = UtilityType.Standard;
-		[SerializeField]public UnderType m_ActiveUnder = UnderType.Standard;
-		
 	}
 	
 	[System.Serializable]
