+++ Assets/Asset Landing Zone/UCC/DragonSlayerItemCollection.asset
@@ -27,11 +27,13 @@ MonoBehaviour:
   m_Enabled: 1
   m_EditorHideFlags: 0
   m_Script: {fileID: 11500000, guid: 4e74c32144565264eb86dc7514aca7f7, type: 3}
-  m_Name: MasterBlaster Ammo
+  m_Name: Battery Power
   m_EditorClassIdentifier: 
   m_ID: 5
   m_CategoryIDs: 00000000
-  m_Description: Consumable type for MasterBlaster, can be one or two batteries
+  m_Description: 'Consumable type for Energy Weapons
+
+'
   m_Capacity: 3.4028235e+38
   m_DroppedItemTypes: []
 --- !u!114 &11400000
