+++ Assets/Opsive/UltimateCharacterController/Demo/Materials/Items/Shooter/Shotgun/MuzzleFlash.mat
@@ -77,4 +77,4 @@ Material:
     - _Color: {r: 1, g: 1, b: 1, a: 1}
     - _EmisColor: {r: 0.19999996, g: 0.19999996, b: 0.19999996, a: 0}
     - _EmissionColor: {r: 0, g: 0, b: 0, a: 1}
-    - _TintColor: {r: 1, g: 1, b: 1, a: 0}
+    - _TintColor: {r: 0, g: 0, b: 0, a: 0}
